refactor examples to share a module
